basic kexec tools for ppc64 (compile warning cleanups)
Hello,
	The basic_kexec_tools_for_ppc64 had a few issues with variable
redeclarations that was throwing some compile warnings on GCC 4, though I
was not seeing some of those on my GCC ver 3 on SLES9. This is a revised
patch, which cleans up the errors and this should compile ok on ppc64. I also
cleaned up a few other compile warnings that were showing up on my system.
	Thanks to Michael Neuling for pointing this out.

Thanks and Regards,
Sharada

This provides the basic functionality in kexec-elf-ppc64.c to load the various
segments required for kexec-tools on ppc64. It also adds ppc64 arch support
in build scripts and cleans up some misc error in member deferencing to allow
the build to succeed.

Signed-off-by: R Sharada <sharada@in.ibm.com>
Signed-off-by: Mohan Kumar <mohan@in.ibm.com>
Signed-off-by: Maneesh Soni <maneesh@in.ibm.com>