- kexec man page
I posted this man page on May 11.  It was pointed out to me that the
something had trashed the nroff formating.
This is a reposting of the man page.  (not a patch, just the man page).

Thanks
--

David Wilder
IBM Linux Technology Center
Beaverton, Oregon, USA
dwilder@us.ibm.com
(503)578-3789

Signed-off-by: Maneesh Soni <maneesh@in.ibm.com>