ppc64 ignore args-linux
This patches changes argument parsing for ppc64 for the command "kexec -p".
We now ignore the args-linux option.  This allows for more consistency
of the command line between archs.  Intel platforms
use the --args-linux option ppc64 dose not.  With this change the same
command-line can used now on both system types.

Signed-off-by: David Wilder <dwilder@us.ibm.com>
Signed-off-by: Maneesh Soni <maneesh@in.ibm.com>