x86_64: Setup CS when booting vmlinux
The purgatory code in kexec-tools does not currently setup CS when booting a
64-bit ELF file such as a vmlinux file. This together with the fact that the
Linux kernel doesn't reload CS properly if booted from the 64-bit entry point
means that booting a vmlinux may fail under certain conditions.

The only known combination that triggers this problem is when kexec-tools and
kexec are used to load a x86_64 vmlinux under a dom0 Linux running under the
Xen hypervisor.

This patch is needed for sure to reload kernels with version <= 2.6.17. There
are fixes for this problem in the URL below, but if a fix will be included in
2.6.18 or not is unknown at this time.

http://permalink.gmane.org/gmane.linux.kernel/438998

Signed-off-by: Magnus Damm <magnus@valinux.co.jp>

Removed some trailing whitespace

Signed-off-by: Simon Horman <horms@verge.net.au>