kexec-tools: --noio option to disable I/O in purgatory code.
SN platform support PIO in a different way to generic IA64 platform. It
does not support most of the legacy I/O ports.

Give an --noio option to kexec-tools to disable I/O in purgatory code.

This patch also removed an unused io.h in kexec-tools.

Signed-off-by: Zou Nan hai <nanhai.zou@intel.com>

Edited to consistently use tabs instead of spaces for intentation,
remove one instance of trailing whitespace, and fix indentation
of noio line in options[].

Signed-off-by: Simon Horman <horms@verge.net.au>