build-no-partial-rebuild-on-clean
[BUILD] Don't do a partial build on clean

$(OBJDIR)/kexec/purgatory.c is a target and needs special treatment,
else calling make clean on an already clean tree will cause it -
and as a result a whole bunch of other stuff - to be built only
to be immediately cleaned.

There are some more pathological cases, where users manually
removed .d files, and then these are regenerated (but not removed)
on make clean. However this change does ensure that the following
is a bit more sensible.

  make clean; make clean

Signed-OFf-By: Simon Horman <horms@verge.net.au>