kexec-ktools: somewhat less nested code in patch_efi_memmap()
This patch reduces the nesting in patch_efi_memmap() by
jumping to the next interation of the inner for loop
if the following condition is true.

if (seg->start < mstart || seg->start >= mend)

This is instead of a reasonably large ammount of code inside the if
conditional if the converse is true.

This makes things somewhat easier to read as the nesting is already
quite deep, and many lines do not fit easily within 80 columns.

Signed-Off-By: Simon Horman <horms@verge.net.au>