kexec-tool: i386: Only load known segments
This was a warning flagged by the compiler, that loadaddr
may be used uninitilised. I'm not sure if this happens in practice,
but it is possible, if the seg type is NBI_SEG_NEGATIVE, or
some other undefined value.

It seems best to print a warning and skip to the next segment in
this case. But I'm not sure if this is correct.

Signed-Off-By: Simon Horman <horms@verge.net.au>