ia64-vmm
This adds vmm support to kexec-tool for ia64. This is annalogous
to the same feature that is present in the latest version of elilo.
It is a method of booting a vmm (hypervisor) such as Xen.
Essentially it works as follows.

* If the --vmm argument is not provided, then the kernel is booted as normal,
  no changes
* Else, the image specified by --vmm is placed loaded into the elf
  segments, where the Linux kernel image would otherwise have gone.
  And the Linux kernel image, allong with its length is loaded into a
  seprate segment, and passed as new entry at the end of the boot parameters.
  This is somewhat similar to how initramfs/initramd images are passed
  to a booting kernel, and can work in conjunction with that feature.
  On boot (or in this case on kexec) the vmm (hypervisor) will be
  loaded instead of a Linux kernel, and the hypervisor will then load up
  the Linux kernel as it sees fit.

This is needed in order for kexec from Xen to Xen, using the
port of kexec to Xen that I am working on, to work.

I am not entirely fond of this design, and i think that developing
an ia64 variant of multiboot would be much nicer. However it is
an existing method that is currently in widespread use through
its incarntation in elilo. And if multiboot is added in future,
it can be done as a separate boot method, and thus orthogonal to this
patch.

In order to use this code a number of other changes are needed,
in particular:

  1. Xen and the corresponding Linux Kernel needs to be patched with
     the port of kexec to ia64-xen that I have been working on.
     I will post the latest version of these patches to xen-devel
     shortly.

  2. The currently hardcoded PAGE_OFFSET value in purgatory needs
     to be changed from the Linux value to the Xen value. I will
     post a very hackish, definately not to be released, patch
     after this patch which includes a comment that explains this problem
     more clearly.

Also, xen->linux and linux->xen is still very much work in progress due
to the problem described at the following link
http://permalink.gmane.org/gmane.linux.ports.ia64/14995

However, from an infastructure point of view I think it would be good to
apply this code, so that kexec-tool is one step closer to being able to
support vmms (hypervisors). The patch does not alter any existing
behaviour, it just adds a new feature. Bugs asside, the only real danger
seems to be confusion for end-users, perhaps we could comment out the help
text to hide the feature from the lay user, or attach a big fat warning to it.

Signed-Off-By: Simon Horman <horms@verge.net.au>