kexec-tools: Make use of crash_create_elf64_headers() (ia64) V2
kexec-tools: Make use of crash_create_elf64_headers() (ia64) V2

Use the new function provided by crashdump-elf.c instead of duplicating the
same functionality in prepare_crash_memory_elf64_headers().

This patch also makes the code create a separate kernel program header - just
like x86_64 does. This may or may not be correct, but making the architectures
behave as similar as possible is probably a good thing.

On top of that the patch contains a bugfix for the fact that space for crash
note program headers are never allocated...

Version 2 of this patch removes alignment and the callback from elf_info.
This together with a cosmetic crashdum -> crashdump fix. I'm not 100% sure
if the removal of the alignment is correct though.

Signed-off-by: Magnus Damm <magnus@valinux.co.jp>

Removed trailing whitespace

Signed-off-by: Simon Horman <horms@verge.net.au>