kexec-tools: Introduce crashdump-xen.c and Xen support V2
kexec-tools: Introduce crashdump-xen.c and Xen support V2

This patch adds the new file crashdump-xen.c that implements Xen support. The
Xen support is not complete yet in the sense that a special program header
for the hypervisor isn't created. Crash notes for physical cpus are created
so basic support is at least provided by this patch.

Version 2 of this patch includes a cleaner implementation for crashdump-elf.c
together with a bugfix for xen_get_nr_phys_cpus().

Signed-off-by: Magnus Damm <magnus@valinux.co.jp>

Removed trailing whitespace

Signed-off-by: Simon Horman <horms@verge.net.au>