send slave cpus to SAL slave loop on crash (IA64)
On Tue, Nov 21, 2006 at 07:13:56AM +0800, Zou Nan hai wrote:
> This patch make normal "kexec -l" first try physical address suggested
> by vmlinux. 
> 
> If there is no enough memory, kexec tools will search /proc/iomem and
> find a place to put the new kernel.
> 
> This is necessary for "kexec -l" to work on SN platform.
> 
> Signed-off-by: Zou Nan hai <nanhai.zou@intel.com>
> 

Hi Nan hai, 

sorry for not responding sooner.

I have take a look at your patch with an eye to including it into 
kexec-tools-testing. It does seem like a good feature.

Below is an updated version of the patch that cleans up your
implementation a little. In particular.

* update_loaded_segments() returns -1 on error, 0 on success
  - amongst other things the previous incarnation could return nothing
    at all in once case.
* else { if () { ; } } -> else if () { ; }
* < 80 columns wide
* removed trailing whitespace

Is it ok to apply this incarntation of the patch?
I can make my changes a second patch if you prefer.

Signed-off-by: Simon Horman <horms@verge.net.au>

* Incremental version of above patch
* Include kexec-ia64.h in kexec-elf-ia64.c for update_loaded_segments()

Signed-off-by: Simon Horman <horms@verge.net.au>