kexec-tools: don't rely on errno to detect errors
Errno is designed to check for additional information if
errors occur. Not to be some kind of global flag set by either
libc or applications to flag that an error occured.

This patch removes the checking of errnor for some prior error somewhere.
It relies on some earlier patches which check functions like lseek()
for errors.

Signed-off-by: Simon Horman <horms@verge.net.au>

Tested and working fine. Ack.

Acked-by: Sachin P. Sant <sachinp@in.ibm.com>