IA64 kexec-tools: memory_ranges arrays scalability issue
There are two memory_ranges arrays declared in the IA64 kexec code.
One in kexec-ia64.c and the other one in crashdump-ia64.c.

They currently were allocated as a hard-coded size of array. Since
SN systems may scale to hunders of nodes and each node may contain
up to 12 DIMMs, the hard-coded size of 1024 is not enough for very
large systems.

The size of either array can not be greater than the number of
entries in /proc/iomem, which is saved as "max_memory_ranges"
and is used to dynamically allocate the two arrays.

Signed-off-by:  Jay Lan <jlan@sgi.com>
Signed-off-by:  Simon Horman <horms@verge.net.au>