Fix wrong max_memory_ranges value
The value of max_memory_ranges is calculated in count_memory_ranges() and only
counts the memory@ nodes. However, in get_devtree_details(), pci@ nodes are also
added to exclude_range just at the end of the loop.

This patch changes count_memory_ranges() to also consider pci@ nodes in
/proc/device-tree.

Signed-off-by: Bernhard Walle <bwalle@suse.de>
Signed-off-by: Simon Horman <horms@verge.net.au>