Move memory range variables
Move memory range variables

The common kexec code is currently using two global variables to keep
track of memory ranges. Other data is kept in a per-instance structure.
This mix is of per-instance and global variables is confusing and leads
to messy code in general. So let's not.

This patch moves the global variables into struct kexec_info and makes
sure that structure is passed along where needed. No logic is changed.

Signed-off-by: Magnus Damm <magnus@valinux.co.jp>
Signed-off-by: Simon Horman <horms@verge.net.au>