kexec-tools: ppc64: use kernels slave loop for purgatory
Purgatory doesn't really care about the SMP cpus.  But if we leave them
behind, they end up getting lost when the kernel overwrites purgatory or
the previous kernel.  The current slave handling in purgatory doesn't
have any handshaking to make sure the cpus have moved on before leaving.

Instead of moving the slave cpus up to purgatory and then back down to
the kernel, just copy bytes 4-255 from the kernel and use it as the
purgatory start / slave hold block.


Signed-off-by: Milton Miller <miltonm@bga.com>
Signed-off-by: Simon Horman <horms@verge.net.au>