Allow architectures to specify the location of /proc/iomem
On ia64 XEN it is sometimes neccessary to use an alternate location for the
information that is usually provided by /proc/iomem.  By having the path
returned by a function, which can be overriden on a per-architecture basis,
it is possible for ia64 XEN to make use of existing generic code. Hopefully
other achitectures can use this infastructure as the need arises.

If the machine parameter is zero, then iomem file relating to the currently
running kernel should be returned. If the machine parameter is non-zero
then iomem file, relating to the underlying hypervisor, should be returned.
In the simple case, these will be the same file.

Signed-off-by: Simon Horman <horms@verge.net.au>

 kexec/arch/i386/crashdump-x86.c      |    4 ++--
 kexec/arch/i386/kexec-x86.c          |    2 +-
 kexec/arch/i386/x86-linux-setup.c    |    2 +-
 kexec/arch/ia64/crashdump-ia64.c     |    9 +++++----
 kexec/arch/ia64/kexec-elf-ia64.c     |    2 +-
 kexec/arch/ia64/kexec-ia64.c         |    5 +++--
 kexec/arch/ppc/kexec-ppc.c           |    2 +-
 kexec/arch/s390/kexec-s390.c         |    5 +++--
 kexec/arch/x86_64/crashdump-x86_64.c |   10 +++++-----
 kexec/arch/x86_64/kexec-x86_64.c     |    2 +-
 kexec/crashdump-xen.c                |    6 ++++--
 kexec/kexec-iomem.c                  |   31 +++++++++++++++++++++++--------
 kexec/kexec.h                        |    6 +++---
 13 files changed, 53 insertions(+), 33 deletions(-)

9079040b40f643cfc9eb3d425dffa0ca8fd573e1