Use zlib if present
Hi,

This fixes a minor regression that occured with Jeremy's recent
reworking of the build system. I guess he wasn't dealing with
compressed images and thus didn't notice that zlib support had
been disabled.

The fix involves updating an #ifdef to use the new symbol,
and ensuring that -lz is passed to the linker as neccessary.

Cc: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Simon Horman <horms@verge.net.au>