kexec-tools: [ia64] always use /proc/iomem_machine on ia64 xen
When I first added proc_iomem() it was for ia64 Xen as
certain regions are added /proc/iomem_machine instead of /proc/iomem
on that platform.

On further examination I have found that all regions need to be
added to /proc/iomem_machine.

This patch changes the ia64 implementation of proc_iomem()
to always return /proc/iomem_machine if xen is in use.

A subsequent patch will remove the machine parameter to proc_iomem() which
is no longer of any use.

Signed-off-by: Simon Horman <horms@verge.net.au>