kexec-tools: mips: support big-endian mips (repost)
[ Reposted with correct linux-mips address ]

Hi,

this patch switches the mips support in kexec-tools around a little bit.
All the files and directories containing "mipsel" have been renamed
to contain "mips" instead.

This is kind of consistent with the way that ARCH=mips in the kernel
works for both big and little endian.

After a small amount of tweaking, which is also included in this patch, the
code compiles and works fine for big endian mips as well as small endian
mips. All you need to do is compile using an appropriate compiler.

That is to say, kexec-tools's build system doesn't need to
be told about which endienness the code is being compiled for.

I have added kept mipsel as a supported "architecture" via ./configure,
though its just an alias for mips now. This is consistent with how
other architectures such as sh are treated. But I'm happy to remove
mipsel from ./configure if the mips people want that.

I tested this patch using qemu and the 2.6.24.3 tag of the mips-2.6 git
tree compiled for the qemu machine type for both big and little endian.
The qemu machine type has subsequently been removed, and kexec-tools
needs some work in order to function with qemu - as far as I understand
the way the boot parameters are passed needs to be fixed, likely
in purgatory. However, this is not related to the changes
introduced in this patch.

I intend to merge this patch into kexec-tools-testing if
no alarm bells are sounded.

Signed-off-by: Simon Horman <horms@verge.net.au>