Allow building for ppc32 platforms.
Currently any ppc32 compile will error out with "Please, fix this for
your platform".  While this is a great way to encourage people to update
the code it seesm to be a disinsentive for distros to package
kexec-tools for powerpc (incl. 64-bit).

Make get_memory_ranges() do the right thing in the GameCube case and
fail at runtime for all other ppc32 cases.

Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>
Signed-off-by: Simon Horman <horms@verge.net.au>