Document kexec without -e/-l/-f/-p and clarify -f
This patch tries to clarify the manpage in two aspects:

It describes the meaning of kexec without an -e / -l / -f / -p option (i.e.
only with kernel, initrd and command line).

It also documents the difference between -f and normal kexec call better.

I made the patch to clarify the manpage because there was some confusion in
https://bugzilla.novell.com/show_bug.cgi?id=387601.

Signed-off-by: Bernhard Walle <bwalle@suse.de>
Signed-off-by: Simon Horman <horms@verge.net.au>