Don't use /sys/firmware/memmap for Xen
On Xen, we have to use /proc/iomem to retrieve the memory area for the kexec'd
kernel, not /sys/firmware/memmap. Dom0 kernel gets a E820 map that contains
only one region:

    0000000000000000-0000000018e5e000 (System RAM)

Compared to the /proc/iomem:

    00000000-0009cbff : System RAM
    0009cc00-0009ffff : reserved
    000ce000-000d3fff : reserved
    000e0000-000fffff : reserved
    00100000-1fd6ffff : System RAM
      01000000-04ffffff : Crash kernel
      1ec00000-1fbfffff : Hypervisor code and data
        1f0b4680-1f0b4873 : Crash note
        1f0b4900-1f0b4a93 : Crash note
        1f0b4b80-1f0b4d13 : Crash note
        1f0b4e00-1f0b4f93 : Crash note
    ...

Without that patch, /proc/vmcore is empty in the kexec'd kernel and I'm unable
to copy the crashdump.

Signed-off-by: Bernhard Walle <bwalle@suse.de>
Signed-off-by: Simon Horman <horms@verge.net.au>